+++ test/offline/offline_test_main.cpp
@@ -60,6 +60,8 @@ const std::unordered_set<std::string> g_functions_expected_to_reject = {
     // display
     "vblank-handler",  // asm
     "vif1-handler", "vif1-handler-debug",
+    // stats-h
+    "(method 11 perf-stat)", "(method 12 perf-stat)",
 
     // sync-info
     "(method 15 sync-info)",         // needs *res-static-buf*
@@ -73,8 +75,6 @@ const std::unordered_set<std::string> g_functions_to_skip_compiling = {
     "abs", "ash", "min", "max", "lognor",
     // weird PS2 specific debug registers:
     "breakpoint-range-set!",
-    // int128 fancy stuff.
-    "(method 3 vec4s)", "(method 2 vec4s)",
     // does weird stuff with the type system.
     "print", "printl", "inspect",
     // inline assembly
@@ -116,9 +116,6 @@ const std::unordered_set<std::string> g_functions_to_skip_compiling = {
 
     // dma-disasm
     "disasm-dma-list",  // missing a single cast :(
-
-    // gs
-    "(method 3 gif-tag)",  // inspect for a 128-bit type.
 
     // math camera
     "transform-point-vector!", "transform-point-qword!", "transform-point-vector-scale!",